Implemented bearer token authentication on each route

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,13 +15,14 @@ export class TasksService {
     ) { }
 
 
-    public async getTask(filterDto: GetTaskFilterDto): Promise<Task[]> {
+    public async getTask(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
         // return await this.taskRepository.find();
-        return this.taskRepository.getTask(filterDto);
+        return this.taskRepository.getTask(filterDto, user);
     }
 
-    public async getTaskByID(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    public async getTaskByID(id: number, user: User): Promise<Task> {
+        
+        const found = await this.taskRepository.findOne({ where: {id, userId: user.id}});
 
         if (!found) {
             throw new NotFoundException(`Task with ID ${id} not found`);
@@ -30,20 +32,20 @@ export class TasksService {
     }
 
 
-    public async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    public async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    public async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    public async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         } 
     }
 
-    public async updateTask(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskByID(id);
+    public async updateTask(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskByID(id, user);
         task.status = status;
         return await task.save();
     }
